Add area note editing to settings
